Generalized assistant with config.ini
