+++ assistant/config.ini
@@ -1,0 +1,23 @@
+; It's likely that this file will only be used for initial startup purposes
+; Other items such as user data and permissions can be read directly in the "assistant.py" file
+[ID]
+Name = KAREN
+Purpose = to assist SPARC members
+Location = SPARC Lab
+First_Run = FALSE ; Not yet implemented
+
+[Keys]
+DialogFlow = d8cd9faa2fe14731b1187d05b7d6f409 ; Example - not yet implemented
+
+[Preferences]
+Input = Voice    ; Can be Text, Voice, or Slack
+Output = Voice ; Can be Text, Voice, or Slack
+Language = English
+
+[UserData] ; May be moved to different file
+User_Name = Josh ; Not yet implemented
+Admin_Pass = SPARK 3 6 7 ; Not yet implemented
+
+[Permissions] ; May be moved to different file
+Location = TRUE ; Not yet implemented
+Active_Recording = TRUE ; Not yet implemented
