+++ assistant/voice_client.py
@@ -4,7 +4,7 @@ Organization: Student Projects and Research Committee at Auburn University
 Project: Lab Assistant
 Description: A voice based version of Karen.  Uses speech to text and text to speech to interact with user.
 """
-from voice_properties import *
+from properties import *
 import manage_suites
 
 
@@ -25,5 +25,13 @@ if __name__ == "__main__":
     mic = -1
     while mic not in range(0, 10):
         mic = int(raw_input("< "))
-    karen = Assistant("Karen", "English", "to assist SPARC members", "SPARC Lab", mic)
+    config = ConfigParser.ConfigParser()
+    config.read("config.ini")
+    name = config.get('ID', 'Name')
+    language = config.get('Preferences', 'Language')
+    purpose = config.get('ID', 'Purpose')
+    set_location = config.get('ID', 'Location')
+    input = config.get('Preferences', 'Input')
+    output = config.get('Preferences', 'Output')
+    karen = Assistant(name, language, purpose, set_location, input, output, mic)
     main(karen)
